added so many different transformations
added transformations on the work page as well as the contact gear has
some whistles on it, the transform on the main page has been cleaned up
as well. the work page gear is slightly broken,
major issues with the work page, I’m not sure if the sliding main is
possible with only css, i will probably have a dummy fix of just <a>
tags linking to certain parts of the page.
check out some of the really broken looking code pens to see more